Add a iterative parsing and summary class.
Large traces don't fit in memory so iterating over the trace and letting
analyses extract the information they need is more effective. The in
memory trace can still be generated if that is wanted.

Instead of just printing summary the data, store it. This will be useful
when it comes to comparing traces from each probe invocation.